move dev menu to devTools folder
