+++ src/app/App.tsx
@@ -2,12 +2,12 @@ import { GoogleSignin } from "@react-native-google-signin/google-signin"
 import { GlobalStore } from "app/store/GlobalStore"
 import { codePushOptions } from "app/system/codepush"
 import { AsyncStorageDevtools } from "app/system/devTools/AsyncStorageDevTools"
+import { DevMenuWrapper } from "app/system/devTools/DevMenu/DevMenuWrapper"
 import { setupFlipper } from "app/system/devTools/flipper"
 import { useRageShakeDevMenu } from "app/system/devTools/useRageShakeDevMenu"
 import { useErrorReporting } from "app/system/errorReporting/hooks"
 import { ModalStack } from "app/system/navigation/ModalStack"
 import { usePurgeCacheOnAppUpdate } from "app/system/relay/usePurgeCacheOnAppUpdate"
-import { DevMenuWrapper } from "app/utils/DevMenu/DevMenuWrapper"
 import { useDevToggle } from "app/utils/hooks/useDevToggle"
 import { addTrackingProvider } from "app/utils/track"
 import {
