+++ src/app/Scenes/Onboarding/Onboarding.tsx
@@ -7,8 +7,8 @@ import {
 import { FPSCounter } from "app/Components/FPSCounter"
 import { OAuthProvider } from "app/store/AuthModel"
 import { GlobalStore } from "app/store/GlobalStore"
+import { DevMenu as DevMenuDefault } from "app/system/devTools/DevMenu/DevMenu"
 import { ArtsyKeyboardAvoidingViewContext } from "app/utils/ArtsyKeyboardAvoidingView"
-import { DevMenu as DevMenuDefault } from "app/utils/DevMenu/DevMenu"
 import { NetworkAwareProvider } from "app/utils/NetworkAwareProvider"
 import { useDevToggle } from "app/utils/hooks/useDevToggle"
 import { Platform, View } from "react-native"
