+++ src/app/system/devTools/DevMenu/DevMenu.tsx
@@ -23,10 +23,10 @@ import { GlobalStore } from "app/store/GlobalStore"
 import { EnvironmentKey, environment } from "app/store/config/EnvironmentModel"
 import { DevToggleName, FeatureName, devToggles, features } from "app/store/config/features"
 import { Versions } from "app/store/migration"
+import { CodePushOptions } from "app/system/devTools/DevMenu/CodePushOptions"
 import { eigenSentryReleaseName } from "app/system/errorReporting//sentrySetup"
 import { dismissModal, goBack, navigate } from "app/system/navigation/navigate"
 import { RelayCache } from "app/system/relay/RelayCache"
-import { CodePushOptions } from "app/utils/DevMenu/CodePushOptions"
 import { useUnleashEnvironment } from "app/utils/experiments/hooks"
 import { useBackHandler } from "app/utils/hooks/useBackHandler"
 import { capitalize, compact, sortBy } from "lodash"
